// src/lib/openapi/adminapi.schema.ts
export interface components {
    schemas: {
        /** CreateEntityIndividual */
        CreateEntityIndividual: {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            entity_type: 'individual';
            /**
             * Format: email
             * @example random@example.com
             */
            email: string;
            gender: components['schemas']['Gender'];
            /** @example John */
            first_name: string;
            /** @example Doe */
            last_name: string;
            /** @example 53 Sungei Tengah Rd */
            street: string;
            /** @example Singapore */
            city: string;
            /** @example 698998 */
            postal_code: string;
            country: components['schemas']['CountryCode2'];
            nationality: components['schemas']['CountryCode2'];
            birth_country: components['schemas']['CountryCode2'];
            birth_date: components['schemas']['Date'];
            /** @example Paris */
            birth_place: string;
        };
// ...
    };
// ...
}